Say "encode", not "serialize"; move the dev entry to index.mjs
Docs: every prose use of serialize/serializing/serialization becomes
encode, across the README and both references. The `serializer` field of
the ReScript `S.transform` API stays -- it's a real field name (S.res:438),
and renaming it in the docs would document API that doesn't exist. Two
adjacent fixes fell out of the sentence that defined the term: it called
the library "The ReScript Schema", and it glossed the reverse operation as
"decoding", which now names a different operation.

Entry: esbuild's output is index.mjs everywhere, not just in the artifact.
`src/S.res` compiles to a file named after itself, so an S.mjs beside it
was one rescript.json "suffix" change away from being silently
overwritten -- the same hazard the artifact already avoids. src/S.d.mts
moves to index.d.mts, and the spec harness, bundleSize, introspect, the
fuzzer, four test files and CLAUDE.md follow the rename.

The fuzzer builds its baseline with the ref's own pack.ts, so revisions
before this commit still emit src/S.mjs; it now resolves whichever exists.
Verified against HEAD: 1320 comparisons, 0 diffs.

bundleSize.yaml is regenerated rather than hand-edited. Only its header
line moves -- total stays 23504, since the bundle is byte-identical.

Two guards added while the rename was in flight:

- pack.ts strips src/ by allowlist (.res, .res.mjs, .res.js, .d.ts) rather
  than by extension. src/ is copied wholesale and partly gitignored, so a
  stale build output in a dirty tree used to ride along -- which is exactly
  what happened here with the pre-rename src/S.mjs.
- The packed README must match the repo root one. They're byte-identical
  duplicates with nothing syncing them, and npm renders the packed copy, so
  a root-only edit would publish stale text.
