+++ packages/spec/harness.ts
@@ -2,7 +2,7 @@
 // executing the real schema.
 //
 // Unlike format.ts (which runs on published sury), this half imports the
-// in-development sury SOURCE (`../sury/src/S.mjs`), because goldens must reflect
+// in-development sury SOURCE (`../sury/index.mjs`), because goldens must reflect
 // the code under test — that's how `spec check` catches codegen changes.
 //
 // There is no code-generation step: packages/sury/tests/spec_test.ts loops
@@ -15,7 +15,7 @@ import { fileURLToPath } from "node:url";
 import { Document, parse as parseYaml, parseDocument, stringify as stringifyYaml, isMap, isSeq } from "yaml";
 import { diffLinesUnified } from "@vitest/utils/diff";
 import ts from "typescript";
-import * as S from "../sury/src/S.mjs";
+import * as S from "../sury/index.mjs";
 import {
   KEY_ORDER,
   TS_KEY_ORDER,
@@ -819,7 +819,7 @@ export const checkSpec = async (
 // Part of the compared text, so an edited or dropped header reads as stale
 // like any other drift.
 const BUNDLE_SIZE_HEADER = [
-  "# Minified+gzipped bytes per public export of src/S.mjs, plus `total` for the whole entry.",
+  "# Minified+gzipped bytes per public export of index.mjs, plus `total` for the whole entry.",
   "# Generated by `pnpm spec check --write` — every row is measured, so never hand-write one.",
 ].join("\n");
