+++ packages/sury/index.d.mts
@@ -1,0 +1,2 @@
+// Types for the generated index.mjs (TS resolves .mjs imports to .d.mts).
+export * from "./src/S.js";
