+++ packages/sury/jsr.json
@@ -3,5 +3,5 @@
   "version": "11.0.0-rc.0",
   "license": "MIT",
   "exclude": ["!src", "!rescript.json", "!README.md", "!package.json", "!docs"],
-  "exports": "./src/S.mjs"
+  "exports": "./index.mjs"
 }
