+++ packages/sury/scripts/unionFuzz.ts
@@ -17,7 +17,7 @@
 // Generation is seeded, so a reported diff reproduces from its seed alone.
 
 import { execFileSync } from "node:child_process";
-import { mkdtempSync, rmSync, symlinkSync } from "node:fs";
+import { existsSync, mkdtempSync, rmSync, symlinkSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 
@@ -44,6 +44,13 @@ const build = (cwd: string): void => {
 // A worktree has no node_modules of its own; the bundler and its deps are
 // identical for both sides, so borrowing the main install is both correct and
 // much faster than a second install.
+// The ref tree builds with its own scripts/pack.ts, and revisions before the
+// entry was renamed emit src/S.mjs instead.
+const entryPath = (tree: string): string => {
+  const current = join(tree, "packages/sury/index.mjs");
+  return existsSync(current) ? current : join(tree, "packages/sury/src/S.mjs");
+};
+
 const checkout = (ref: string): string => {
   const dir = mkdtempSync(join(tmpdir(), "sury-fuzz-"));
   const tree = join(dir, "tree");
@@ -262,8 +269,8 @@ const main = async (): Promise<void> => {
   let currentModule: Sury;
   try {
     build(repoRoot);
-    refModule = await import(join(refTree, "packages/sury/src/S.mjs"));
-    currentModule = await import(join(repoRoot, "packages/sury/src/S.mjs"));
+    refModule = await import(entryPath(refTree));
+    currentModule = await import(entryPath(repoRoot));
 
     const next = rng(seed);
     const pick = <T,>(list: readonly T[]): T =>
