+++ packages/sury/specs/bundleSize.yaml
@@ -1,4 +1,4 @@
-# Minified+gzipped bytes per public export of src/S.mjs, plus `total` for the whole entry.
+# Minified+gzipped bytes per public export of index.mjs, plus `total` for the whole entry.
 # Generated by `pnpm spec check --write` — every row is measured, so never hand-write one.
 total: 23504
 exports:
