--- packages/sury/src/S.d.mts
@@ -1,2 +1,0 @@
-// Types for the generated src/S.mjs (TS resolves .mjs imports to .d.mts).
-export * from "./S.js";
