+++ packages/sury/tests/S_test.ts
@@ -1,6 +1,6 @@
 import { test, expectTypeOf, assertType } from "vitest";
 
-import * as S from "../src/S.mjs";
+import * as S from "../index.mjs";
 
 // FIXME: S.max should be applied to output
 // From https://x.com/dzakh_dev/status/1963982551208309222
