+++ packages/sury/tests/union_planner_regression_test.ts
@@ -1,6 +1,6 @@
 import { test } from "vitest";
 
-import * as S from "../src/S.mjs";
+import * as S from "../index.mjs";
 
 test("union semantic discriminators use identity and SameValueZero", (t) => {
   const firstSymbol = Symbol("same-description");
